+++ here-naksha-lib-core/src/main/java/com/here/naksha/lib/core/AbstractTask.java
@@ -613,6 +613,9 @@ public abstract class AbstractTask<RESULT, SELF extends AbstractTask<RESULT, SEL
         break;
       }
       // Failed, conflict, repeat
+      log.info(
+          "Concurrency conflict while incrementing instance level threadCount from {}. Will retry...",
+          threadCount);
     }
   }
 
@@ -649,6 +652,9 @@ public abstract class AbstractTask<RESULT, SELF extends AbstractTask<RESULT, SEL
       if (counter == null) {
         Long existing = actorUsageMap.putIfAbsent(actorId, 1L);
         if (existing != null) {
+          log.info(
+              "Concurrency conflict while initializing threadCount to 1 for actorId [{}]. Will retry...",
+              actorId);
           continue; // Repeat, conflict with other thread
         }
         return;
@@ -669,6 +675,10 @@ public abstract class AbstractTask<RESULT, SELF extends AbstractTask<RESULT, SEL
         break;
       }
       // Failed, conflict, repeat
+      log.info(
+          "Concurrency conflict while incrementing actor level threadCount from {} for actorId [{}]. Will retry...",
+          counter,
+          actorId);
     }
   }
 
@@ -693,6 +703,9 @@ public abstract class AbstractTask<RESULT, SELF extends AbstractTask<RESULT, SEL
           log.error("Invalid actor usage value for actor: " + actorId + " value: " + current);
         }
         if (!actorUsageMap.remove(actorId, current)) {
+          log.info(
+              "Concurrency conflict while removing actor level threadCount for actorId [{}]. Will retry...",
+              actorId);
           continue;
         }
         break;
@@ -700,6 +713,10 @@ public abstract class AbstractTask<RESULT, SELF extends AbstractTask<RESULT, SEL
         break;
       }
       // Failed, repeat, conflict with other thread
+      log.info(
+          "Concurrency conflict while decrementing actor level threadCount from {} for actorId [{}]. Will retry...",
+          current,
+          actorId);
     }
   }
 }