+++ here-naksha-lib-core/src/main/java/com/here/naksha/lib/core/NakshaContext.java
@@ -221,6 +221,7 @@ public final class NakshaContext {
         return newValue;
       }
       // Conflict, two threads seem to want to update the same key the same time!
+      logger.info("Concurrency conflict while updating attachment map for key {}", valueClass);
     }
   }