+++ here-naksha-lib-core/src/main/java/com/here/naksha/lib/core/util/json/JsonMap.java
@@ -42,6 +42,8 @@ import java.util.function.Function;
 import org.jetbrains.annotations.ApiStatus.AvailableSince;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * A map that uses {@link String} key and arbitrary values. The map is thread safe for concurrent
@@ -54,6 +56,8 @@ import org.jetbrains.annotations.Nullable;
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
 public class JsonMap
     implements Map<@NotNull String, @Nullable Object>, Iterable<Map.Entry<@NotNull String, @Nullable Object>> {
+
+  private static final Logger logger = LoggerFactory.getLogger(JsonMap.class);
 
   /**
    * Create a new empty map.
@@ -278,6 +282,7 @@ public class JsonMap
       final Object result =
           FibMap.put(key, oldValue, newValue, true, rootMutable(), this::intern, this::conflict);
       if (result instanceof FibMapConflict) {
+        logger.info("Concurrency conflict while setting value for key {}. Will retry...", key);
         continue;
       }
       assert result == oldValue;
@@ -305,6 +310,7 @@ public class JsonMap
       final Object result =
           FibMap.put(key, original, newValue, true, rootMutable(), this::intern, this::conflict);
       if (result instanceof FibMapConflict) {
+        logger.info("Concurrency conflict while setting value for key {}. Will retry...", key);
         continue;
       }
       assert result == original;
@@ -339,6 +345,7 @@ public class JsonMap
         final Object result =
             FibMap.put(key, original, UNDEFINED, true, rootMutable(), this::intern, this::conflict);
         if (result instanceof FibMapConflict) {
+          logger.info("Concurrency conflict while removing value for key {}. Will retry...", key);
           continue;
         }
         assert result == UNDEFINED;
@@ -350,6 +357,7 @@ public class JsonMap
       final Object result =
           FibMap.put(key, original, newValue, true, rootMutable(), this::intern, this::conflict);
       if (result instanceof FibMapConflict) {
+        logger.info("Concurrency conflict while setting value for key {}. Will retry...", key);
         continue;
       }
       assert result == original;
@@ -468,6 +476,7 @@ public class JsonMap
         SIZE.getAndAdd(this, -oldSize);
         return;
       }
+      logger.info("Concurrency conflict while clearing map. Will retry...");
     }
   }